requirements.txt
aiogram==3.17.0
discord.py==2.4.0
Requests==2.32.3
